requirements.txt
google-cloud-datastore==2.1.6